requirements.txt
flask
pyserial
pynmea2
